requirements.txt
serpapi
requests
python-dotenv
schedule
tldextract